(fix) it might be that images are in double type

+++ thesis_code/training/evaluation/vectorize_test_dataset.py
@@ -47,7 +47,7 @@ def main():
         data_i = mri_i.get_fdata()  # type: ignore
 
         # Get MRI vectorizer output
-        inp = torch.from_numpy(data_i).unsqueeze(0).unsqueeze(0).to(args.device)
+        inp = torch.from_numpy(data_i).float().unsqueeze(0).unsqueeze(0).to(args.device)
         x = mri_vectorizer(inp)
         x = x.detach().cpu().squeeze(0).squeeze(0)
         mri_vectorizer_out[i] = x
